partial fix
